Add Feedback table and update Discriminator column
Introduce migration 20250130011311_AddFeedbackTable to add a new Feedback table with Id, FoundResource, Comments, and SubmittedAt columns. Update the Discriminator column in the People table to have a max length of 21 characters. Reflect these changes in ApplicationDbContextModelSnapshot and update ProductVersion to 8.0.11. Remove explicit seed values in UseIdentityColumns method calls. Update BuildTargetModel in Designer file and implement Up and Down methods for the migration.