Add IsReviewed property to Feedback and FeedbackViewModel
Introduced a new boolean property `IsReviewed` in the `Feedback`
class within the `PC2.Models` namespace to indicate whether the
feedback has been reviewed by an administrator. The property is
initialized to `false` by default. Similarly, added the same
`IsReviewed` property to the `FeedbackViewModel` class to track
the review status in the view model as well. These changes
enhance the feedback management functionality.