Refactor registration endpoint routing (#373)
* Refactor registration endpoint routing

Updated the handling of requests to the "/Identity/Account/Register" endpoint by replacing `endpoints.MapGet` and `endpoints.MapPost` within an `app.UseEndpoints` block with direct calls to `app.MapGet` and `app.MapPost`. This change simplifies the routing configuration and ensures users are redirected to the "/Identity/Account/Login" page when accessing the registration page.

* Refactor registration route handling

Updated the "/Identity/Account/Register" route to use an asynchronous method for redirecting users to the login page. This change replaces the previous implementation that utilized `Task.Factory.StartNew`, simplifying the code and improving readability and efficiency.