Update footer change Twitter icon to X (#426)
Updated the Twitter link in the footer to point to X.com and replaced the Twitter icon with the new X.com logo using an external SVG file. Adjusted the X.com logo's styling (size, padding, and background) for consistency. Increased the Facebook icon size for better visual alignment. Added the "x-logo.svg" file to the project.